Move message out of asynchronous
