+++ pymongo/synchronous/mongo_client.py
@@ -1667,13 +1667,13 @@ class MongoClient(common.BaseObject, Generic[_DocumentType]):
         if operation.conn_mgr:
             server = self._select_server(
                 operation.read_preference,
-                operation.session,
+                operation.session,  # type: ignore[arg-type]
                 operation.name,
                 address=address,
             )
 
             with operation.conn_mgr._alock:
-                with _MongoClientErrorHandler(self, server, operation.session) as err_handler:
+                with _MongoClientErrorHandler(self, server, operation.session) as err_handler:  # type: ignore[arg-type]
                     err_handler.contribute_socket(operation.conn_mgr.conn)
                     return server.run_operation(
                         operation.conn_mgr.conn,
@@ -1703,7 +1703,7 @@ class MongoClient(common.BaseObject, Generic[_DocumentType]):
         return self._retryable_read(
             _cmd,
             operation.read_preference,
-            operation.session,
+            operation.session,  # type: ignore[arg-type]
             address=address,
             retryable=isinstance(operation, _Query),
             operation=operation.name,
