+++ test/__init__.py
@@ -45,12 +45,11 @@ from urllib.parse import quote_plus
 import pymongo
 import pymongo.errors
 from bson.son import SON
-from pymongo import common
+from pymongo import common, message
 from pymongo.common import partition_node
 from pymongo.hello_compat import HelloCompat
 from pymongo.server_api import ServerApi
 from pymongo.ssl_support import HAVE_SSL, _ssl  # type:ignore[attr-defined]
-from pymongo.synchronous import message
 from pymongo.synchronous.database import Database
 from pymongo.synchronous.mongo_client import MongoClient
 from pymongo.uri_parser import parse_uri
