+++ test/auth_oidc/test_auth_oidc.py
@@ -36,14 +36,14 @@ from pymongo._azure_helpers import _get_azure_response
 from pymongo._gcp_helpers import _get_gcp_response
 from pymongo.cursor_shared import CursorType
 from pymongo.errors import AutoReconnect, ConfigurationError, OperationFailure
+from pymongo.hello_compat import HelloCompat
 from pymongo.operations import InsertOne
 from pymongo.synchronous.auth_oidc import (
     OIDCCallback,
     OIDCCallbackContext,
     OIDCCallbackResult,
 )
-from pymongo.synchronous.hello_compat import HelloCompat
-from pymongo.synchronous.uri_parser import parse_uri
+from pymongo.uri_parser import parse_uri
 
 ROOT = Path(__file__).parent.parent.resolve()
 TEST_PATH = ROOT / "auth" / "unified"
