+++ test/test_client.py
@@ -106,7 +106,6 @@ from pymongo.read_preferences import ReadPreference
 from pymongo.server_description import ServerDescription
 from pymongo.server_selectors import readable_server_selector, writable_server_selector
 from pymongo.server_type import SERVER_TYPE
-from pymongo.synchronous import message as message_old
 from pymongo.synchronous.command_cursor import CommandCursor
 from pymongo.synchronous.cursor import Cursor, CursorType
 from pymongo.synchronous.database import Database
@@ -1458,7 +1457,7 @@ class TestClient(IntegrationTest):
         with self.assertRaises(AutoReconnect):
             client = rs_client(connect=False, serverSelectionTimeoutMS=100)
             client._run_operation(
-                operation=message_old._GetMore(
+                operation=message._GetMore(
                     "pymongo_test",
                     "collection",
                     101,
